Refactor: Unify Canal and St. Venant Models
This commit refactors the canal-related physical objects into a single `UnifiedCanal` class.

The key changes are:
- The logic from the deprecated `Canal` class (a simple integral model) and the `StVenantReach` class (a full hydrodynamic model) have been merged into `UnifiedCanal`.
- A `model_type` parameter in `UnifiedCanal` now controls which model is used (e.g., 'integral', 'st_venant').
- The `NetworkSolver` has been updated to work with `UnifiedCanal` objects that have `model_type='st_venant'`.
- The old `Canal` and `StVenantReach` class files have been deleted.
- All examples, tests, configurations, and documentation across the codebase have been updated to use the new `UnifiedCanal` class, ensuring consistency and preventing breakage.
