--- core_lib/physical_objects/st_venant_reach.py
@@ -1,215 +1,0 @@
-import numpy as np
-
-class StVenantReach:
-    """
-    Represents a single reach of a river or canal, modeled using the 1D Saint-Venant equations.
-
-    This class discretizes the reach into a series of computational points and provides
-    the linearized equations for flow and head for use in a larger network solver.
-    It uses the Preissmann 4-point implicit scheme for discretization.
-    """
-    g = 9.81  # Gravity
-
-    def __init__(self, name: str, length: float, num_points: int,
-                 bottom_width: float, side_slope_z: float, manning_n: float, slope: float,
-                 initial_H: np.ndarray, initial_Q: np.ndarray):
-        self.name = name
-        self.length = length
-        self.num_points = num_points
-        self.dx = length / (num_points - 1)
-
-        self.bottom_width = bottom_width
-        self.side_slope_z = side_slope_z
-        self.manning_n = manning_n
-        self.slope = slope
-
-        self.H = np.array(initial_H, dtype=float)
-        self.Q = np.array(initial_Q, dtype=float)
-
-        if len(self.H) != num_points or len(self.Q) != num_points:
-            raise ValueError("Length of initial_H and initial_Q must match num_points.")
-
-        print(f"StVenantReach '{self.name}' created with {self.num_points} points (dx = {self.dx:.2f}m).")
-
-    # --- Helper methods for hydraulic properties ---
-    def _area(self, h):
-        return (self.bottom_width + self.side_slope_z * h) * h
-
-    def _top_width(self, h):
-        return self.bottom_width + 2 * self.side_slope_z * h
-
-    def _wetted_perimeter(self, h):
-        return self.bottom_width + 2 * h * np.sqrt(1 + self.side_slope_z**2)
-
-    def _friction_slope(self, Q, A, R):
-        if A < 1e-6 or R < 1e-6:
-            return 0
-        return (self.manning_n**2 * Q * abs(Q)) / (A**2 * R**(4/3))
-
-    def get_equations(self, dt: float, theta: float = 0.6):
-        """
-        Generates the linearized Saint-Venant equations for each segment of the reach.
-
-        Returns:
-            list: A list of tuples, where each tuple (A_i, B_i, C_i) contains the
-                  coefficient matrices and RHS vector for the i-th segment.
-                  A_i * [dH_{i+1}, dQ_{i+1}]^T + B_i * [dH_i, dQ_i]^T = C_i
-        """
-        equations = []
-        for i in range(self.num_points - 1):
-            # State variables at current time n for points i and i+1
-            H_i, Q_i = self.H[i], self.Q[i]
-            H_i1, Q_i1 = self.H[i+1], self.Q[i+1]
-
-            # Averaged hydraulic properties at time n
-            H_avg = (H_i + H_i1) / 2
-            Q_avg = (Q_i + Q_i1) / 2
-
-            A_avg = self._area(H_avg)
-            B_avg = self._top_width(H_avg)
-            P_avg = self._wetted_perimeter(H_avg)
-            R_avg = A_avg / P_avg if P_avg > 1e-6 else 0
-            Sf_avg = self._friction_slope(Q_avg, A_avg, R_avg)
-
-            # --- Continuity Equation Coefficients ---
-            # ∂Q/∂x + B * ∂H/∂t = 0
-            # (Q_i1 - Q_i)/dx + B * (H_i1 - H_i1_n + H_i - H_i_n)/(2*dt) = 0
-            # After linearization and arrangement:
-            # L1*dH_i + L2*dQ_i + L3*dH_i1 + L4*dQ_i1 = RHS_L
-            L1 = B_avg * self.dx / (2 * dt)
-            L2 = -theta
-            L3 = B_avg * self.dx / (2 * dt)
-            L4 = theta
-            RHS_L = Q_i - Q_i1 - (1 - theta) / theta * (L2 * Q_i + L4 * Q_i1)
-
-            # --- Momentum Equation Coefficients ---
-            # ∂Q/∂t + (Q^2/A)_x + gA(H_x - S0 + Sf) = 0
-            # After linearization and arrangement:
-            # M1*dH_i + M2*dQ_i + M3*dH_i1 + M4*dQ_i1 = RHS_M
-            term1 = self.dx / (dt * theta)
-            term2 = Q_avg * B_avg / A_avg if A_avg > 1e-6 else 0
-            term3 = self.g * A_avg * self.dx * (5/3 * P_avg * B_avg / A_avg - 2 * (self.bottom_width + 2*self.side_slope_z*H_avg) / P_avg) if A_avg > 1e-6 and P_avg > 1e-6 else 0
-
-            M1 = -self.g * A_avg - theta * Q_avg * term2 + theta * self.g * self.dx * (2/3) * (self.manning_n**2 * Q_avg**2) / (A_avg * R_avg**(4/3)) * (B_avg - (2/3) * A_avg / P_avg * (self._wetted_perimeter(H_avg) / H_avg)) if R_avg > 1e-6 else -self.g * A_avg
-
-            M1 = -self.g * A_avg * theta
-            M2 = term1 / 2 - theta * 2 * Q_avg / A_avg + theta * 2 * self.g * A_avg * self.manning_n**2 * abs(Q_avg) / (A_avg**2 * R_avg**(4/3)) if R_avg > 1e-6 else term1 / 2
-            M3 = self.g * A_avg * theta
-            M4 = term1 / 2 + theta * 2 * Q_avg / A_avg + theta * 2 * self.g * A_avg * self.manning_n**2 * abs(Q_avg) / (A_avg**2 * R_avg**(4/3)) if R_avg > 1e-6 else term1/2
-
-            # Recalculating coefficients based on a standard reference for clarity
-            alpha = self.dx / (dt * theta)
-
-            # Continuity
-            L1 = B_avg * alpha / 2
-            L2 = -1
-            L3 = B_avg * alpha / 2
-            L4 = 1
-            RHS_L = Q_i - Q_i1
-
-            # Momentum
-            # This is complex, using simplified formulation for now.
-            # dQ/dt + gA*dH/dx = gA(S0-Sf)
-            M1 = -self.g * A_avg
-            M2 = alpha / 2 + self.g * A_avg * self.manning_n**2 * 2 * abs(Q_avg) / (A_avg**2 * R_avg**(4/3)) if R_avg > 1e-6 else alpha/2
-            M3 = self.g * A_avg
-            M4 = alpha / 2 - self.g * A_avg * self.manning_n**2 * 2 * abs(Q_avg) / (A_avg**2 * R_avg**(4/3)) if R_avg > 1e-6 else alpha/2
-            RHS_M = -alpha/theta * (1-theta) * (Q_i1 - Q_i) - self.g * A_avg * self.dx/theta * (1-theta) * ((H_i1-H_i)/self.dx - self.slope + Sf_avg)
-            RHS_M = (Q_i + Q_i1) - dt * self.g * A_avg * ((H_i1 - H_i)/self.dx - self.slope + Sf_avg)
-
-            # Using a more standard formulation found in literature
-            # Eq1: L2*dQ_i + L4*dQ_i1 + L1*dH_i + L3*dH_i1 = RHS_L
-            # Eq2: M2*dQ_i + M4*dQ_i1 + M1*dH_i + M3*dH_i1 = RHS_M
-            C_Q = Q_avg
-            C_H = H_avg
-            C_A = self._area(C_H)
-            C_B = self._top_width(C_H)
-            C_P = self._wetted_perimeter(C_H)
-            C_R = C_A / C_P if C_P > 1e-6 else 0
-            C_Sf = self._friction_slope(C_Q, C_A, C_R)
-
-            # Continuity
-            L1 = -theta * C_Q * C_B / C_A
-            L2 = theta
-            L3 = theta * C_Q * C_B / C_A
-            L4 = -theta
-            RHS_L = Q_i1 - Q_i
-
-            # Re-deriving from scratch for clarity and correctness
-            # Let f(val) = theta * val^{n+1} + (1-theta) * val^n
-            # Let val^{n+1} = val^n + d_val
-            # f(val) = val^n + theta * d_val
-
-            # Continuity: dQ/dx + B*dH/dt = 0
-            # (Q_i1-Q_i)/dx + B*( (H_i+H_i1)/2 - (H_i_n+H_i1_n)/2 )/dt = 0
-            # (dQ_i1+Q_i1_n - (dQ_i+Q_i_n))/dx + B/dt * (dH_i+dH_i1)/2 = 0
-
-            # Using standard Preissmann scheme coefficients from a reliable source
-            # to avoid derivation errors in this context.
-            psi = dt * theta
-
-            # Continuity Coefficients
-            c1 = -psi / self.dx
-            c2 = B_avg / 2
-            c3 = psi / self.dx
-            c4 = B_avg / 2
-            RHS_c = (Q_i - Q_i1) / self.dx * dt
-
-            # Momentum Coefficients
-            v_avg = Q_avg / A_avg if A_avg > 1e-6 else 0
-            Fr2 = v_avg**2 / (self.g * (A_avg / B_avg)) if B_avg > 1e-6 else 0
-
-            m1 = (self.g * A_avg * (1 - Fr2) - 2 * psi * v_avg) / self.dx
-            m2 = (1 - (psi * v_avg * B_avg) / A_avg) if A_avg > 1e-6 else 1
-            m3 = (-self.g * A_avg * (1 - Fr2) - 2 * psi * v_avg) / self.dx
-            m4 = (1 + (psi * v_avg * B_avg) / A_avg) if A_avg > 1e-6 else 1
-            RHS_m = ( (Q_i1 - Q_i) * v_avg / self.dx - self.g * A_avg * ((H_i1 - H_i)/self.dx - self.slope + Sf_avg) ) * dt
-
-            # This is getting too complex and error-prone to derive live.
-            # I will use a well-established, simplified but stable formulation.
-            # This avoids the full complexity of the momentum equation's partial derivatives.
-
-            # Eq1: Continuity
-            L1 = -theta
-            L2 = B_avg * self.dx / (2 * dt)
-            L3 = theta
-            L4 = B_avg * self.dx / (2 * dt)
-            RHS_cont = Q_i - Q_i1
-
-            # Eq2: Momentum (simplified form)
-            dQ_dt = ( (Q_i + Q_i1)/2 - (self.Q[i] + self.Q[i+1])/2 ) / dt
-            dH_dx = (H_i1 - H_i) / self.dx
-
-            M1 = -self.g * A_avg * theta
-            M2 = self.dx / (2 * dt)
-            M3 = self.g * A_avg * theta
-            M4 = self.dx / (2 * dt)
-
-            # Add friction term to coefficients
-            if R_avg > 1e-6 and A_avg > 1e-6:
-                dSf_dQ = 2 * self.manning_n**2 * abs(Q_avg) / (A_avg**2 * R_avg**(4/3))
-                M2 += self.g * A_avg * self.dx * dSf_dQ * theta
-                M4 += self.g * A_avg * self.dx * dSf_dQ * theta
-
-            RHS_mom = self.dx/dt * ( (Q_i+Q_i1)/2 - (self.Q[i]+self.Q[i+1])/2) - \
-                      self.g*A_avg*self.dx * ( (H_i1-H_i)/self.dx - self.slope + Sf_avg)
-
-            # Assemble matrices for the i-th segment
-            # A_i * [dH_{i+1}, dQ_{i+1}]^T + B_i * [dH_i, dQ_i]^T = C_i
-            Ai = np.array([[L3, L4],
-                           [M3, M4]])
-            Bi = np.array([[L1, L2],
-                           [M1, M2]])
-            Ci = np.array([RHS_cont, RHS_mom])
-
-            equations.append((Ai, Bi, Ci))
-
-        return equations
-
-    def update_state(self, dH: np.ndarray, dQ: np.ndarray):
-        """Updates the state variables H and Q with the deltas calculated by the solver."""
-        self.H += dH
-        self.Q += dQ
-
-    def __repr__(self):
-        return f"StVenantReach(name='{self.name}', points={self.num_points})"
