+++ core_lib/physical_objects/unified_canal.py
@@ -9,12 +9,14 @@ class UnifiedCanal(PhysicalObjectInterface):
     A unified model for a canal reach that can represent several different
     simplified models based on a `model_type` parameter.
     """
+    g = 9.81  # Gravity
+
     def __init__(self, name: str, initial_state: State, parameters: Parameters, **kwargs):
         super().__init__(name, initial_state, parameters)
 
         self.model_type = self._params.get('model_type', 'integral_delay')
 
-        # Common parameters
+        # Common parameters for simple models
         self._state['water_level'] = initial_state.get('water_level', 5.0)
         self._state['inflow'] = initial_state.get('inflow', 0.0)
         self._state['outflow'] = initial_state.get('outflow', 0.0)
@@ -34,6 +36,26 @@ class UnifiedCanal(PhysicalObjectInterface):
             self.storage_constant = self._params.get('storage_constant', 1200)
             self.level_storage_ratio = self._params.get('level_storage_ratio', 0.005)
             self.storage = self._state.get('water_level', 0.0) / self.level_storage_ratio
+        elif self.model_type == 'st_venant':
+            self.length = self._params['length']
+            self.num_points = self._params['num_points']
+            self.dx = self.length / (self.num_points - 1)
+
+            self.bottom_width = self._params['bottom_width']
+            self.side_slope_z = self._params['side_slope_z']
+            self.manning_n = self._params['manning_n']
+            self.slope = self._params['slope']
+
+            initial_H = self._params.get('initial_H', np.full(self.num_points, 5.0))
+            initial_Q = self._params.get('initial_Q', np.full(self.num_points, 10.0))
+
+            self.H = np.array(initial_H, dtype=float)
+            self.Q = np.array(initial_Q, dtype=float)
+
+            if len(self.H) != self.num_points or len(self.Q) != self.num_points:
+                raise ValueError("Length of initial_H and initial_Q must match num_points.")
+
+            print(f"UnifiedCanal '{self.name}' (st_venant model) created with {self.num_points} points (dx = {self.dx:.2f}m).")
 
         # History buffer for delay models
         self.inflow_history = None
@@ -51,6 +73,8 @@ class UnifiedCanal(PhysicalObjectInterface):
             self._step_integral_delay_zero(dt)
         elif self.model_type == 'linear_reservoir':
             self._step_linear_reservoir(dt)
+        elif self.model_type == 'st_venant':
+            raise NotImplementedError("The 'st_venant' model cannot be run with step(). It must be used with the NetworkSolver.")
         else:
             raise ValueError(f"Unknown canal model type: {self.model_type}")
 
@@ -108,3 +132,80 @@ class UnifiedCanal(PhysicalObjectInterface):
     @property
     def is_stateful(self) -> bool:
         return True
+
+    # --- St. Venant Model Methods ---
+
+    def _area(self, h):
+        return (self.bottom_width + self.side_slope_z * h) * h
+
+    def _top_width(self, h):
+        return self.bottom_width + 2 * self.side_slope_z * h
+
+    def _wetted_perimeter(self, h):
+        return self.bottom_width + 2 * h * np.sqrt(1 + self.side_slope_z**2)
+
+    def _friction_slope(self, Q, A, R):
+        if A < 1e-6 or R < 1e-6:
+            return 0
+        return (self.manning_n**2 * Q * abs(Q)) / (A**2 * R**(4/3))
+
+    def get_equations(self, dt: float, theta: float = 0.6):
+        """
+        Generates the linearized Saint-Venant equations for each segment of the reach.
+        This method is only applicable when model_type is 'st_venant'.
+        """
+        if self.model_type != 'st_venant':
+            raise RuntimeError("get_equations() is only valid for the 'st_venant' model type.")
+
+        equations = []
+        for i in range(self.num_points - 1):
+            H_i, Q_i = self.H[i], self.Q[i]
+            H_i1, Q_i1 = self.H[i+1], self.Q[i+1]
+            H_avg = (H_i + H_i1) / 2
+            Q_avg = (Q_i + Q_i1) / 2
+            A_avg = self._area(H_avg)
+            B_avg = self._top_width(H_avg)
+            P_avg = self._wetted_perimeter(H_avg)
+            R_avg = A_avg / P_avg if P_avg > 1e-6 else 0
+            Sf_avg = self._friction_slope(Q_avg, A_avg, R_avg)
+
+            # Simplified but stable Preissmann scheme formulation
+            # Eq1: Continuity
+            L1 = -theta
+            L2 = B_avg * self.dx / (2 * dt)
+            L3 = theta
+            L4 = B_avg * self.dx / (2 * dt)
+            RHS_cont = Q_i - Q_i1
+
+            # Eq2: Momentum
+            M1 = -self.g * A_avg * theta
+            M2 = self.dx / (2 * dt)
+            M3 = self.g * A_avg * theta
+            M4 = self.dx / (2 * dt)
+
+            if R_avg > 1e-6 and A_avg > 1e-6:
+                dSf_dQ = 2 * self.manning_n**2 * abs(Q_avg) / (A_avg**2 * R_avg**(4/3))
+                M2 += self.g * A_avg * self.dx * dSf_dQ * theta
+                M4 += self.g * A_avg * self.dx * dSf_dQ * theta
+
+            RHS_mom = self.dx/dt * ( (Q_i+Q_i1)/2 - (self.Q[i]+self.Q[i+1])/2) - \
+                      self.g*A_avg*self.dx * ( (H_i1-H_i)/self.dx - self.slope + Sf_avg)
+
+            Ai = np.array([[L3, L4], [M3, M4]])
+            Bi = np.array([[L1, L2], [M1, M2]])
+            Ci = np.array([RHS_cont, RHS_mom])
+
+            equations.append((Ai, Bi, Ci))
+
+        return equations
+
+    def update_state(self, dH: np.ndarray, dQ: np.ndarray):
+        """
+        Updates the state variables H and Q with the deltas calculated by the solver.
+        This method is only applicable when model_type is 'st_venant'.
+        """
+        if self.model_type != 'st_venant':
+            raise RuntimeError("update_state() is only valid for the 'st_venant' model type.")
+
+        self.H += dH
+        self.Q += dQ
