--- docs/water_system/02_physical_models/07_canal.md
@@ -1,71 +1,0 @@
-# 物理模型: 统一渠池 (UnifiedCanal)
-
-> [!WARNING]
-> **模型已更新**: 从 v2.0 开始，旧的渠池模型 (`Canal`, `IntegralDelayCanal`, `RiverChannel`) 已被弃用，并统一由 `UnifiedCanal` 模型代替。请更新您的代码以使用新模型。
-
-本篇文档介绍用于模拟明渠或渠池水力学行为的统一物理模型 `UnifiedCanal`。
-
-*   **源代码**: `core_lib/physical_objects/unified_canal.py`
-
-## 概述
-
-`UnifiedCanal` 是一个灵活的、统一的渠池模型。它通过一个 `model_type` 参数，可以在多种不同复杂度的简化模型之间切换，以适应从精细化水力学仿真到高层次控制策略设计的不同需求。
-
-这种设计大大提高了代码的可维护性和扩展性，允许用户在同一框架下轻松比较和切换不同的数学模型。
-
-### 核心参数
-
--   `model_type` (str): 指定要使用的模型类型。必须是以下之一：
-    -   `'integral'`: 简单的积分（水库）模型。
-    -   `'integral_delay'`: 积分加时滞模型。
-    -   `'integral_delay_zero'`: 积分、时滞加零点模型。
-    -   `'linear_reservoir'`: 线性水库模型。
-
----
-
-## 模型详解
-
-### 1. 积分模型 (`model_type='integral'`)
-
-这是最基础的模型，将渠池视为一个简单的水库。
-
-*   **核心思想**: 渠池的水量变化等于净流入量（入流减去出流）。
-*   **适用场景**: 当渠道的动态特性可以忽略不计，或只需要进行水量平衡计算时。
-*   **参数**:
-    -   `surface_area` (float): 渠池的水面面积 (m²)。
-    -   `outlet_coefficient` (float): 出流系数，用于根据水位计算出流 `q_out = C * sqrt(h)`。
-*   **工作机制**: 在每个时间步，模型根据当前水位计算出流量，然后根据入流量和出流量的差值，使用 `volume += (inflow - outflow) * dt` 来更新蓄水量，并最终更新水位。
-
-### 2. 积分-时滞模型 (`model_type='integral_delay'`)
-
-该模型在积分模型的基础上，增加了时滞效应，是长距离输水渠道的经典简化模型。
-
-*   **核心思想**: 上游的流量变化需要经过一段时间的传播才能影响到下游。
-*   **适用场景**: 控制系统设计，特别是当输送延迟是主要考虑因素时。
-*   **参数**:
-    -   `gain` (float): 过程增益 (K)，通常与水面面积成反比 (`K = 1 / A_surface`)。
-    -   `delay` (float): 水流从上游到下游的传播时间 (s)。
-*   **工作机制**: 模型内部维护一个历史入流的队列。在每个时间步，它将 `t - delay` 时刻的入流量作为当前的有效入流，并以此来更新水位。出流量被直接定义为这个延迟后的入流量。
-
-### 3. 积分-时滞-零点模型 (`model_type='integral_delay_zero'`)
-
-这是对积分-时滞模型的进一步改进，通过增加一个“零点”项来更好地逼近真实渠道的初始响应。
-
-*   **核心思想**: 增加一个与入流变化率相关的项，可以更准确地模拟波的传播效应。
-*   **适用场景**: 需要更精确控制响应的系统设计。
-*   **参数**:
-    -   `gain` (float): 过程增益 (K)。
-    -   `delay` (float): 传播延迟 (s)。
-    -   `zero_time_constant` (float): 零点时间常数 (Tz)。
-*   **工作机制**: 出流量不仅取决于延迟后的入流量，还取决于延迟入流量的变化率（导数），即 `q_out(t) = q_in(t - τ) + Tz * d/dt(q_in(t - τ))`。
-
-### 4. 线性水库模型 (`model_type='linear_reservoir'`)
-
-该模型常用于模拟河流或渠道的汇流过程，它能捕捉到一定的水量扩散效应。
-
-*   **核心思想**: 渠段的蓄水量与出流量成线性关系 (`S = k * q_out`)。
-*   **适用场景**: 流域汇流计算、洪水演进的简化模拟。
-*   **参数**:
-    -   `storage_constant` (float): 蓄水常数 (k)，反映了渠段的调蓄能力。
-    -   `level_storage_ratio` (float): 水位-蓄量转换系数，用于从蓄水量估算水位。
-*   **工作机制**: 模型求解常微分方程 `dS/dt = q_in - q_out`。在每个时间步，根据离散化的方程 `q_out_new = (k*q_out_old + dt*q_in) / (k + dt)` 来计算新的出流量，并更新总蓄水量和水位。
