+++ examples/canal_model/canal_model_comparison/README.md
@@ -12,7 +12,7 @@
 
 比较的四种渠道模型是：
 
-1.  **积分模型 (`Canal`)**: 这是最简单的模型，将渠道表示为单个水库，其中水位是净流量（流入量 - 流出量）的积分。它不考虑波的传播时间。其控制方程为：
+1.  **积分模型 (`UnifiedCanal(model_type='integral')`)**: 这是最简单的模型，将渠道表示为单个水库，其中水位是净流量（流入量 - 流出量）的积分。它不考虑波的传播时间。其控制方程为：
     `dh/dt = (1/A) * (q_in - q_out)`
     其中 `h` 是水位，`A` 是表面积，`q_in` 是流入量，`q_out` 是流出量。
