+++ mission/example_1/README.zh-CN.md
@@ -6,7 +6,7 @@
 
 ### `example_1_1_physical_model.py`
 *   **目标**: 演示纯物理模型（渠池和闸门）的动态行为。
-*   **内容**: 在一个手动的仿真循环中，我们初始化一个渠池（`Canal`）和一个闸门（`Gate`），并观察在给定上游来流和固定闸门开度的情况下，渠池水位如何随时间变化。
+*   **内容**: 在一个手动的仿真循环中，我们初始化一个渠池（`UnifiedCanal(model_type='integral')`）和一个闸门（`Gate`），并观察在给定上游来流和固定闸门开度的情况下，渠池水位如何随时间变化。
 *   **核心概念**: 物理对象建模与仿真。
 
 ### `example_1_2_physical_io_agent.py`
