+++ mission/example_5/README.zh-CN.md
@@ -6,7 +6,7 @@
 
 ### `example_5_1_turbine_gate_simulation.py`
 *   **目标**: 独立演示单个水轮机和泄洪闸的物理行为，特别是下游水位（顶托）的影响。
-*   **内容**: 此脚本使用 `NetworkSolver` 和 `StVenantReach` 对水工建筑物进行更精细的液压仿真。它分别模拟了水轮机和闸门在下游尾水位动态变化时的流量响应，直观地展示了“顶托”效应对过流能力的影响。
+*   **内容**: 此脚本使用 `NetworkSolver` 和 `UnifiedCanal` (配置为 `model_type='st_venant'`) 对水工建筑物进行更精细的液压仿真。它分别模拟了水轮机和闸门在下游尾水位动态变化时的流量响应，直观地展示了“顶托”效应对过流能力的影响。
 *   **核心概念**: 液压建模，圣维南方程，尾水顶托效应。
 
 ### `example_5_2_multi_turbine_grid.py`
